+++ packages/orq-rc/RELEASES.md
@@ -4159,3 +4159,13 @@ Based on:
 - [typescript v3.14.0-rc.16] packages/orq-rc
 ### Releases
 - [NPM v3.14.0-rc.16] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.16 - packages/orq-rc
+
+## 2025-10-08 14:09:20
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.636.1 (2.723.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.14.0-rc.17] packages/orq-rc
+### Releases
+- [NPM v3.14.0-rc.17] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.17 - packages/orq-rc